Port 00-remix.sh to fish shell
Fish shell is known for its incompatibility by design with POSIX shells. Therefore, 00-remix.sh does not run on fish shell.
It might be controversial to do this port because most fish users are able to figure out a workaround themselves without relying on the upstream, but the logic in this script is so complicated that I believe it is better to provide such a port.